gen-ai/tutorials/inference/lm_invoker/lm_invoker_with_system_prompt/pyproject.toml
[project]
name = "lm-invoker-with-system-prompt"
version = "0.0.0"
description = "LM invoker with system prompt example"
requires-python = ">=3.11,<3.14"
readme = "README.md"
dependencies = [
    "gllm-core>=0.3.0,<0.4.0",
    "gllm-inference[openai]>=0.5.0,<0.6.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[[tool.uv.index]]
name = "gen-ai-internal"
url = "https://glsdk.gdplabs.id/gen-ai-internal/simple/"

[tool.uv.sources]
gllm-core = { index = "gen-ai-internal" }
gllm-inference = { index = "gen-ai-internal" }
